stock_trade/requirements.txt
asgiref==3.6.0
Django==4.1.9
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
pymongo==3.12.3
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.2.4
